timlne changes

+++ java/spring/src/main/java/com/angrysurfer/spring/dao/SessionStatus.java
@@ -40,7 +40,7 @@ public class SessionStatus implements Serializable {
     private Float tempoInBPM;
     private Integer loopCount;
     private Integer parts;
-    private Double noteOffset;
+    private Integer noteOffset;
     private Boolean playing;
     private Boolean hasSolos;
